+++ CHANGES.md
@@ -1,7 +1,7 @@
 # Changelog
 
 ## Unreleased Changes
-
+- Add documentation for `deployment_prediction_generation.py` example DAG.
 
 ## 0.3.0
 - Rename example DAGs to product/GTM desired names