+++ CHANGES.md
@@ -2,6 +2,8 @@
 
 ## Unreleased Changes
 - Add documentation for `deployment_prediction_generation.py` example DAG.
+- Add the `CrossValidationMetricsOperator` to compute scoring for all partitions for a model.
+- Add the `ScoreBacktestsModelOperator` to score all backtests for a datetime partitioned model.
 
 ## 0.3.0
 - Rename example DAGs to product/GTM desired names