+++ CHANGES.md
@@ -3,7 +3,7 @@
 ## Unreleased Changes
 - Add the `CrossValidationMetricsOperator` to compute scoring for all partitions for a model.
 - Add the `ScoreBacktestsModelOperator` to score all backtests for a datetime partitioned model.
-
+- Add the `GetFeaturesUsedOperator` to get the features used to train the model.
 
 ## 0.3.0
 - Rename example DAGs to product/GTM desired names